refactor(GraphQLProvider): update GraphQL hooks and queries for improved structure and functionality
