+++ src/nomas/modules/api/graphql/mutations/index.ts
@@ -1,1 +1,2 @@
 export * from "./mutation-verify-message"
+export * from "./mutation-ephemeral-colyseus"
