+++ src/nomas/modules/api/graphql/queries/index.ts
@@ -1,2 +1,3 @@
 export * from "./query-store-items"
+export * from "./query-request-signature"
 export * from "./query-pets"
